Add support for variable definition blocks
This touches a lot of items.  We have to know the machine word size to
answer the question about size of an opaque type.  We need a model of
memory (stack, heap, and static).

Also fixes some bugs with functions (not collected under the correct
nonterminal) and _gotos_ (not saving/restoring labels).