Fix "forward declaration" pass
Subroutine tokens in the `<extrinsics>` need to reference the set of all
extrinsics available at their scope, even ones declared after them.  To
accomplish this, when the "registration" scan occurs, allocate a number
and use that number within the `subroutine` or subroutine-returning`
token.  When the registration scan ends, associate the extrinsic scope
number with the full set of extrinsics available in that scope.