Change the way previous loading of a unit is determined
With the change to supporting multiple "runtimes" to allow macros to run
in a segregated environment, the test for whether a unit has yet been
loaded was broken (which might be K Framework limitation).  Change to
explicitly track (runtime, filepath) pairs to determine if the given
unit has yet been loaded (and avoid loading more than once into the same
runtime).

This fixes the problem with test/import-subr.oro that appeared in commit
5387c2948f0b8.